Validaciones en caso de uso y Header
Correccion de llevar la logica de BO al caso de uso. Header cambia segun el rol y se agrega el nombre del usuario.
Se añadio AdministradoresBO y su interfaz.

+++ Ghosttracks/GhosttracksVentaEnLinea/src/main/java/itson/org/ghosttracksventaenlinea/fachada/VentaEnLineaFachada.java
@@ -1,16 +1,20 @@
 package itson.org.ghosttracksventaenlinea.fachada;
 
+import itson.org.ghosttracks.dtos.AdministradorDTO;
 import itson.org.ghosttracks.dtos.CarritoDTO;
 import itson.org.ghosttracks.dtos.ClienteDTO;
 import itson.org.ghosttracks.dtos.PedidoDTO;
 import itson.org.ghosttracks.dtos.ProductoDTO;
 import itson.org.ghosttracks.enums.EstadoPedidoDTO;
+import itson.org.ghosttracks.negocio.interfaces.IAdministradoresBO;
 import itson.org.ghosttracks.negocio.interfaces.ICarritoBO;
 import itson.org.ghosttracks.negocio.interfaces.IClientesBO;
 import itson.org.ghosttracks.negocio.interfaces.IPedidosBO;
 import itson.org.ghosttracks.negocio.interfaces.IProductosBO;
+import itson.org.ghosttracks.negocio.objetosNegocio.AdministradoresBO;
 import itson.org.ghosttracks.negocio.objetosNegocio.CarritoBO;
 import itson.org.ghosttracks.negocio.objetosNegocio.ClientesBO;
+import itson.org.ghosttracks.negocio.objetosNegocio.Excepciones.NegocioException;
 import itson.org.ghosttracks.negocio.objetosNegocio.PedidosBO;
 import itson.org.ghosttracks.negocio.objetosNegocio.ProductosBO;
 import itson.org.ghosttracksventaenlinea.interfaces.IVentaEnLinea;
@@ -25,12 +29,14 @@ public class VentaEnLineaFachada implements IVentaEnLinea{
     private final IProductosBO productosBO;
     private final IClientesBO clientesBO;
     private final ICarritoBO carritoBO;
+    private final IAdministradoresBO administradorBO;
 
     public VentaEnLineaFachada() {
         this.productosBO = new ProductosBO();
         this.clientesBO = new ClientesBO();
         this.carritoBO = new CarritoBO();
         this.pedidosBO = new PedidosBO();
+        this.administradorBO = new AdministradoresBO();
     }
 
     @Override
@@ -40,31 +46,79 @@ public class VentaEnLineaFachada implements IVentaEnLinea{
 
     @Override
     public ProductoDTO consultarDetalleProducto(Long id) throws Exception {
+        if (id == null || id <= 0) {
+            throw new NegocioException("El ID de producto proporcionado no es válido.");
+        }
         return productosBO.obtenerProductoPorId(id);
     }
     
     @Override
     public ClienteDTO consultarPerfilCliente(Long idCliente) throws Exception {
+        if (idCliente == null) {
+            throw new NegocioException("Administrador no encontrado");
+        
+        }
         return clientesBO.obtenerClientePorId(idCliente);
     }
     
     @Override
+    public AdministradorDTO consultarPerfilAdministrador(Long idEmpleado) throws Exception {
+        if (idEmpleado == null) {
+            throw new NegocioException("Administrador no encontrado");
+        
+        }
+        return administradorBO.obtenerAdministradorPorId(idEmpleado);
+    }
+    
+    @Override
     public CarritoDTO agregarAlCarrito(CarritoDTO carritoActual, ProductoDTO producto, Integer cantidad) throws Exception {
+        if (carritoActual == null) {
+            carritoActual = new CarritoDTO();
+        }
+        
+        if (producto == null || cantidad == null || cantidad <= 0) {
+            throw new NegocioException("Producto o cantidad inválidos al intentar agregar al carrito.");
+        }
+        
+        if (producto.getStock() == null || cantidad > producto.getStock()) {
+            throw new NegocioException("No hay suficiente stock para el producto: " + producto.getNombre());
+        }
+        
         return carritoBO.agregarProducto(carritoActual, producto, cantidad);
     }
     
     @Override
     public CarritoDTO eliminarDelCarrito(CarritoDTO carrito, Long idProducto) throws Exception {
+        if (carrito == null || carrito.getProductos().isEmpty()) {
+            throw new NegocioException("No hay productos en el carrito para eliminar.");
+        }
+         
         return carritoBO.eliminarProducto(carrito, idProducto);
     }
     
     @Override
-    public PedidoDTO confirmarCompra(PedidoDTO pedido) throws Exception {
-        return pedidosBO.generarPedido(pedido);
+    public PedidoDTO confirmarCompra(PedidoDTO pedidoDto) throws Exception {
+        if (pedidoDto == null || pedidoDto.getProductos() == null || pedidoDto.getProductos().isEmpty()) {
+            throw new NegocioException("No se puede procesar un pedido vacío.");
+        }
+        
+        if (pedidoDto.getTotal() == null || pedidoDto.getTotal() == 0) {
+                pedidoDto.calcularTotales(150.0); 
+        }
+        
+        return pedidosBO.generarPedido(pedidoDto);
     }
     
     @Override
         public PedidoDTO actualizarEstadoPedido(Long idPedido, EstadoPedidoDTO nuevoEstado) throws Exception {
+        if (idPedido == null || idPedido <= 0) {
+            throw new NegocioException("El ID del pedido no es válido.");
+        }
+        
+        if (nuevoEstado == null) {
+            throw new NegocioException("Debe proporcionar un estado válido para actualizar.");
+        }
+        
         return pedidosBO.actualizarEstadoPedido(idPedido, nuevoEstado);
     }
